Crate buzz_types

This crate contains common types that are used by multiple other buzz crates

Modules

dev

Contains things are used internally by Buzz across crates that are not meant for user consuption

errors

Contains error types

traits

Contains traits

Structs

BuzzContext

Holds metadata about the incoming HttpRequest that’s being handled.

Headers

HttpRequest

Representation of an HTTP request that’s been parsed.

HttpResponse

Holds the data that will be formatted back into an HTTP response by the server.

Inject

Wrapper that indicates a shared reference should be injected.

InjectMut

Wrapper that indicates an exclusive reference should be injected.

Route

Represents a tree of url routes.

RouteMetadata

Contains preparsed route segments and the HTTP method that corresponds to a handler.

Enums

HttpMethod

Represents an HTTP request method.

HttpStatusCode

Represents the status code of an HTTP response.

SegmentType

Represents the type of a url segment.

Type Aliases

Handler

A wrapper around a user defined route handler.

Middleware

A function that can modify a request before it’s handled by routing