Type Definition buzz_types::Handler

pub type Handler = fn(_: Vec<&str>, _: HashMap<&str, &str>, _: Option<&str>, _: BuzzContext<'_>, _: &DependancyInjection) -> Result<HttpResponse, BuzzError>;

A wrapper around a user defined route handler.

When a user defines a function and decorates it with a macro like get a wrapper function is generated that will call the users function with the correct arguments. That wrapper function will have the type signature of Handler. It takes a Vec of route params a HashMap of query params, an optional &str containing the body, a BuzzContext which contains any other metadata the caller might need and a DependancyInjection which can be used to fetch registered types.